move loadingAnim, selectlistitem and tabsmenu to components
