+++ app/benchmarking/components/benchmarkReportViewer.tsx
@@ -2,12 +2,12 @@
 import { useEffect, useState } from 'react';
 import { downloadHtmlReport } from '@/app/benchmarking/utils/reportDownloader';
 import { Button, ButtonType } from '@/app/components/button';
+import { LoadingAnimation } from '@/app/components/loadingAnimation';
 import { MainSectionSurface } from '@/app/components/mainSectionSurface';
 import { SelectInput } from '@/app/components/selectInput';
 import { colors } from '@/app/customColors';
 import { useGetBenchmarksResultQuery } from '@/app/services/benchmark-api-service';
 import { useGetRunnerByIdQuery } from '@/app/services/runner-api-service';
-import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 import { BenchmarkReport } from './benchmarkReport';
 
 function BenchmarkReportViewer() {
